Cleaning and minor fixes

+++ emBODY/eBcode/arch-arm/embobj/plus/mc/EOaxisController_hid.h
@@ -116,7 +116,7 @@ struct EOaxisController_hid
     uint16_t state_mask;
     
     eOmc_calibration_type_t calibration_type;
-    float pwm_limit_calib;
+    int32_t pwm_limit_calib;
     uint16_t calib_count;
     uint16_t pos_stable;
     int32_t old_pos;
